Extract address-bar page-focus capture/restore into CmuxBrowserOmnibar
Move the omnibar address-bar page-focus capture/restore subsystem out of the
13k-line BrowserPanel.swift god file into a new CmuxBrowserOmnibar package as a
real Repository with a constructor-injected protocol seam.

What moved:
- The three JavaScript snippets (capture, document-start tracking bootstrap,
  restore), copied byte-for-byte.
- The generation-guarded retry state machine (addressBarFocusRestoreGeneration,
  the [0.0, 0.03, 0.09, 0.2] delay schedule, restore-attempt recursion).
- AddressBarPageFocusRestoreStatus and its result/error parser.

New types (in CmuxBrowserOmnibar):
- BrowserOmnibarPageFocusRepository (@MainActor, owns the capture/restore logic
  and the generation counter).
- BrowserOmnibarScriptEvaluating protocol seam (inverts the WKWebView reach).
- AddressBarPageFocusRestoreStatus value enum.

Seams / DI:
- The repository takes a BrowserOmnibarScriptEvaluating and an optional log sink
  via init. The app-side conformer BrowserOmnibarPageFocusAdapter holds the
  BrowserPanel WEAKLY (panel owns repository owns adapter) and always reads the
  panel's current webView at call time, since the panel reassigns its web view
  across navigation/profile switches. This inverts the webView god-reach with no
  retain cycle.

Byte-identical: the JS scripts, retry delays, generation guards, main-queue
asyncAfter scheduling, and debug-log strings are preserved exactly. BrowserPanel
keeps captureAddressBarPageFocusIfNeeded / invalidateAddressBarPageFocusRestoreAttempts
/ restoreAddressBarPageFocusIfNeeded as one-line forwards, so every call site
(BrowserPanel internals + BrowserPanelView) is unchanged. The tracking bootstrap
WKUserScript now reads BrowserOmnibarPageFocusRepository.trackingBootstrapScriptSource.

Left in BrowserPanel (not cleanly separable, would break byte-identical):
the broader omnibar focus-intent state machine (preferredFocusIntent, the
suppress* flags, pendingAddressBarFocusRequestId, requestAddressBarFocus,
setOmnibarVisible) is fused with searchState, the find-field, browserFocusMode,
and AppDelegate.shared in the same @Published class; moving it would require a
cross-domain rewrite. Extracted the largest clean subset.

Tests: 8 Swift Testing cases for the repository (fake evaluator drives capture,
first-attempt restore, terminal no_state/missing_target, generation invalidation
aborting a pending retry, and status mapping).

Est ~284 lines removed from BrowserPanel.swift; budget TSV ratcheted down.
NOTE: full app build validated by CI.
