+++ .github/swift-file-length-budget.tsv
@@ -5,7 +5,7 @@
 17778	Sources/AppDelegate.swift
 16674	Sources/ContentView.swift
 14785	Sources/TerminalController.swift
-13358	Sources/Panels/BrowserPanel.swift
+13074	Sources/Panels/BrowserPanel.swift
 12361	Sources/Workspace.swift
 12093	Sources/GhosttyTerminalView.swift
 12046	cmuxTests/AppDelegateShortcutRoutingTests.swift
